Overhaul core test cases (remove outdated, rename some, move up base functionality, document).
