+++ core/src/test/java/de/jplag/BasicFunctionalityTest.java
@@ -3,15 +3,15 @@ package de.jplag;
 import static org.junit.jupiter.api.Assertions.assertArrayEquals;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 
-import java.io.File;
-import java.util.List;
-import java.util.Optional;
-
 import org.junit.jupiter.api.Test;
 
 import de.jplag.exceptions.ExitException;
 
-class ParallelComparisonTest extends TestBase {
+/**
+ * Test class for the very basic functionality of JPlag.
+ * @author Timur Saglam
+ */
+class BasicFunctionalityTest extends TestBase {
 
     /**
      * The simple duplicate contains obvious plagiarism.
@@ -87,66 +87,4 @@ class ParallelComparisonTest extends TestBase {
         assertEquals(12, biggestMatch.get().matches().size());
     }
 
-    // TODO SH: Methods like this should be moved to the API and also should accept wildcards
-    private double getSelectedPercent(JPlagResult result, String nameA, String nameB) {
-        return getSelectedComparison(result, nameA, nameB).map(JPlagComparison::similarity).orElse(-1.0);
-    }
-
-    private Optional<JPlagComparison> getSelectedComparison(JPlagResult result, String nameA, String nameB) {
-        return result.getAllComparisons().stream()
-                .filter(comparison -> comparison.firstSubmission().getName().equals(nameA) && comparison.secondSubmission().getName().equals(nameB)
-                        || comparison.firstSubmission().getName().equals(nameB) && comparison.secondSubmission().getName().equals(nameA))
-                .findFirst();
-    }
-
-    @Test
-    void testMultiRootDirNoBasecode() throws ExitException {
-        List<String> paths = List.of(getBasePath("basecode"), getBasePath("SimpleDuplicate")); // 3 + 2 submissions.
-        JPlagResult result = runJPlag(paths, it -> it);
-        assertEquals(5, result.getNumberOfSubmissions());
-    }
-
-    @Test
-    void testMultiRootDirSeparateBasecode() throws ExitException {
-        String basecodePath = getBasePath("basecode-base");
-        List<String> paths = List.of(getBasePath("basecode"), getBasePath("SimpleDuplicate")); // 3 + 2 submissions.
-        JPlagResult result = runJPlag(paths, it -> it.withBaseCodeSubmissionDirectory(new File(basecodePath)));
-        assertEquals(5, result.getNumberOfSubmissions());
-    }
-
-    @Test
-    void testMultiRootDirBasecodeInSubmissionDir() throws ExitException {
-        String basecodePath = getBasePath("basecode", "base");
-        List<String> paths = List.of(getBasePath("basecode"), getBasePath("SimpleDuplicate")); // 2 + 2 submissions.
-        JPlagResult result = runJPlag(paths, it -> it.withBaseCodeSubmissionDirectory(new File(basecodePath)));
-        assertEquals(4, result.getNumberOfSubmissions());
-    }
-
-    @Test
-    void testDisjunctNewAndOldRootDirectories() throws ExitException {
-        List<String> newDirectories = List.of(getBasePath("SimpleDuplicate")); // 2 submissions
-        List<String> oldDirectories = List.of(getBasePath("basecode")); // 3 submissions
-        JPlagResult result = runJPlag(newDirectories, oldDirectories, it -> it);
-        int numberOfExpectedComparison = 1 + 3 * 2;
-        assertEquals(numberOfExpectedComparison, result.getAllComparisons().size());
-    }
-
-    @Test
-    void testOverlappingNewAndOldDirectoriesOverlap() throws ExitException {
-        List<String> newDirectories = List.of(getBasePath("SimpleDuplicate")); // 2 submissions
-        List<String> oldDirectories = List.of(getBasePath("SimpleDuplicate"));
-        JPlagResult result = runJPlag(newDirectories, oldDirectories, it -> it);
-        int numberOfExpectedComparison = 1;
-        assertEquals(numberOfExpectedComparison, result.getAllComparisons().size());
-    }
-
-    @Test
-    void testBasecodeInOldDirectory() throws ExitException {
-        String basecodePath = getBasePath("basecode", "base");
-        List<String> newDirectories = List.of(getBasePath("SimpleDuplicate")); // 2 submissions
-        List<String> oldDirectories = List.of(getBasePath("basecode")); // 3 - 1 submissions
-        JPlagResult result = runJPlag(newDirectories, oldDirectories, it -> it.withBaseCodeSubmissionDirectory(new File(basecodePath)));
-        int numberOfExpectedComparison = 1 + 2 * 2;
-        assertEquals(numberOfExpectedComparison, result.getAllComparisons().size());
-    }
 }
