--- core/src/test/java/de/jplag/OptionsTest.java
@@ -1,5 +1,0 @@
-package de.jplag;
-
-public class OptionsTest extends TestBase {
-    // TODO SH: There should be more tests testing with different combinations of JPlagOptions, e.g., base code
-}
